Revert "Develop/2.5 js executor (#2685)"
This reverts commit 1599b24c3a748599a84f877987f452579c997ccf.
