+++ common/queue/src/main/java/org/thingsboard/server/queue/azure/servicebus/TbServiceBusAdmin.java
@@ -18,7 +18,6 @@ package org.thingsboard.server.queue.azure.servicebus;
 import com.microsoft.azure.servicebus.management.ManagementClient;
 import com.microsoft.azure.servicebus.management.QueueDescription;
 import com.microsoft.azure.servicebus.primitives.ConnectionStringBuilder;
-import com.microsoft.azure.servicebus.primitives.MessagingEntityAlreadyExistsException;
 import com.microsoft.azure.servicebus.primitives.ServiceBusException;
 import lombok.extern.slf4j.Slf4j;
 import org.thingsboard.server.queue.TbQueueAdmin;
@@ -72,12 +71,7 @@ public class TbServiceBusAdmin implements TbQueueAdmin {
             client.createQueue(queueDescription);
             queues.add(topic);
         } catch (ServiceBusException | InterruptedException e) {
-            if (e instanceof MessagingEntityAlreadyExistsException) {
-                queues.add(topic);
-                log.info("[{}] queue already exists.", topic);
-            } else {
-                log.error("Failed to create queue: [{}]", topic, e);
-            }
+            log.error("Failed to create queue: [{}]", topic, e);
         }
     }
